prompt-engineering/requirements.txt
prompty==0.1.50
python-dotenv==1.1.0
rich==14.0.0